+++ src/evo_lib/drivers/camera/__init__.py
@@ -1,6 +1,5 @@
 """Camera drivers: UVC capture with built-in ArUco detection and calibration."""
 
-from evo_lib.drivers.camera.aruco import CharucoCalibrationSession
 from evo_lib.drivers.camera.uvc import (
     UvcCamera,
     UvcCameraDefinition,
@@ -9,7 +8,6 @@ from evo_lib.drivers.camera.uvc import (
 )
 
 __all__ = [
-    "CharucoCalibrationSession",
     "UvcCamera",
     "UvcCameraDefinition",
     "UvcCameraVirtual",
