+++ src/evo_lib/drivers/camera/uvc.py
@@ -15,13 +15,6 @@ from evo_lib.driver_definition import (
     DriverInitArgs,
     DriverInitArgsDefinition,
 )
-from evo_lib.drivers.camera.aruco import (
-    CharucoCalibrationSession,
-    _aruco_dict_from_name,
-    _build_charuco_board,
-    _marker_object_points,
-    _now_iso,
-)
 from evo_lib.interfaces.camera import (
     DEFAULT_ARUCO_DICTIONARY,
     DEFAULT_CHARUCO_MARKER_MM,
@@ -32,6 +25,13 @@ from evo_lib.interfaces.camera import (
     Camera,
 )
 from evo_lib.logger import Logger
+from evo_lib.perception.aruco import (
+    CharucoCalibrationSession,
+    _aruco_dict_from_name,
+    _build_charuco_board,
+    _marker_object_points,
+    _now_iso,
+)
 from evo_lib.task import ImmediateResultTask, Task
 
 if TYPE_CHECKING:
