+++ src/evo_lib/perception/__init__.py
@@ -1,0 +1,3 @@
+from evo_lib.perception.aruco import CharucoCalibrationSession
+
+__all__ = ["CharucoCalibrationSession"]
